+++ src/aliases.js
@@ -1,10 +1,6 @@
-import { version as processVersion } from 'process'
-
 import { getProjectVersion } from './project.js'
 
-// `versionRange` can be one of the following aliases:
-//   - `_`: current process's Node.js version
-//   - `.`: current project's Node.js version using `.nvmrc`, etc.
+// `versionRange` can be one of several aliases
 export const resolveAlias = function (versionRange, opts) {
   const getVersion = ALIASES[versionRange]
 
@@ -15,12 +11,16 @@ export const resolveAlias = function (versionRange, opts) {
   return getVersion(opts)
 }
 
-const getCurrentVersion = function () {
-  return processVersion
+const getLatestVersion = function () {
+  return '*'
 }
 
 // List of available aliases
 const ALIASES = {
-  _: getCurrentVersion,
-  '.': getProjectVersion,
+  current: getProjectVersion,
+  // eslint-disable-next-line id-length
+  c: getProjectVersion,
+  latest: getLatestVersion,
+  // eslint-disable-next-line id-length
+  l: getLatestVersion,
 }